+++ R/rmd_vec_mat.R
@@ -20,12 +20,14 @@
 #' @param ps_equal_sign equal sign to be used when a variable name is given
 #' @param ps_env math environment in which matrix should be presented
 #' @param ps_out_type one of the output types available in Rmarkdown
+#' @param ps_out_format output format produced
 #' @return sresult String containing Rmarkdown representation of given vector or matrix
 rmd_vec_mat <- function(pmat,
                         ps_name       = NULL,
                         ps_equal_sign = '=',
                         ps_env        = NULL,
-                        ps_out_type   = 'array') {
+                        ps_out_type   = 'array',
+                        ps_out_format = 'html') {
   # initialise result variables
   sresult <- NULL
   sfooter <- NULL
@@ -39,15 +41,21 @@ rmd_vec_mat <- function(pmat,
       sfooter <- paste0('\\end{', ps_env, '} ', collapse = '')
     }
   }
+  # put content of matrix, determine dimension of matrix
+  ncol <- ncol(pmat)
+  nrow <- nrow(pmat)
+  # bmatrix with more than 10 columns requires setting max column counter,
+  # according to: https://tex.stackexchange.com/questions/3519/how-to-use-more-than-10-tab-stops-in-bmatrix-or-other-amsmath-matrix-environment
+  if (ps_out_type == "bmatrix" && ncol > 10L && ps_out_format == "latex"){
+    sresult <- paste0(sresult, '\\setcounter{MaxMatrixCols}{', ncol, '} ', collapse = '')
+  }
   # name if one is specified
   if (!is.null(ps_name)){
     sresult <- paste0(sresult, ps_name, ' ', ps_equal_sign, ' ', collapse = '')
   }
+
   # put beginning of environment of specifified output type
   sresult <- paste0(sresult, '\\begin{', ps_out_type, '} ', collapse = '')
-  # put content of matrix, determine dimension of matrix
-  ncol <- ncol(pmat)
-  nrow <- nrow(pmat)
   # loop over first to second to last rows
   if (nrow > 1){
     for (idx in 1:(nrow-1)){
@@ -88,6 +96,7 @@ rmd_vec_mat <- function(pmat,
 #' @param ps_name variable name for the matrix
 #' @param ps_equal_sign equal sign to be used when a variable name is given
 #' @param ps_env math environment in which matrix should be presented
+#' @param ps_out_format output format produced
 #' @return String containing bmatrix representation of given matrix
 #'
 #' @examples
@@ -97,7 +106,11 @@ rmd_vec_mat <- function(pmat,
 #' }
 #'
 #' @export bmatrix
-bmatrix <- function(pmat, ps_name = NULL, ps_equal_sign = '=', ps_env = NULL){
+bmatrix <- function(pmat,
+                    ps_name = NULL,
+                    ps_equal_sign = '=',
+                    ps_env = NULL,
+                    ps_out_format = 'html'){
   # check type of pmat
   if (! is.matrix(pmat))
     stop(" *** ERROR: specified pmat is not a matrix")
@@ -106,7 +119,8 @@ bmatrix <- function(pmat, ps_name = NULL, ps_equal_sign = '=', ps_env = NULL){
                      ps_name       = ps_name,
                      ps_equal_sign = ps_equal_sign,
                      ps_env        = ps_env,
-                     ps_out_type   = 'bmatrix'))
+                     ps_out_type   = 'bmatrix',
+                     ps_out_format = ps_out_format))
 }
 
 #'                                                                           #
@@ -121,6 +135,7 @@ bmatrix <- function(pmat, ps_name = NULL, ps_equal_sign = '=', ps_env = NULL){
 #' @param ps_name variable name for the vector
 #' @param ps_equal_sign equal sign to be used when a variable name is given
 #' @param ps_env math environment in which matrix should be presented
+#' @param ps_out_format output format produced
 #' @return String containing bmatrix representation of given vector
 #'
 #' @examples
@@ -130,16 +145,21 @@ bmatrix <- function(pmat, ps_name = NULL, ps_equal_sign = '=', ps_env = NULL){
 #' }
 #'
 #' @export bcolumn_vector
-bcolumn_vector <- function(pvec, ps_name = NULL, ps_equal_sign = '=', ps_env = NULL){
+bcolumn_vector <- function(pvec,
+                           ps_name = NULL,
+                           ps_equal_sign = '=',
+                           ps_env = NULL,
+                           ps_out_format = 'html'){
   # check type of pmat
   if (! is.vector(pvec))
     stop(" *** ERROR: specified pmat is not a vector")
   # Use bmatrix as output type
-  return(rmd_vec_mat(pmat          = as.matrix(pvec, ncol = 1),
+  return(rmd_vec_mat(pmat          = matrix(pvec, ncol = 1),
                      ps_name       = ps_name,
                      ps_equal_sign = ps_equal_sign,
                      ps_env        = ps_env,
-                     ps_out_type   = 'bmatrix'))
+                     ps_out_type   = 'bmatrix',
+                     ps_out_format = ps_out_format))
 }
 
 #'                                                                           #
@@ -154,6 +174,7 @@ bcolumn_vector <- function(pvec, ps_name = NULL, ps_equal_sign = '=', ps_env = N
 #' @param ps_name variable name for the vector
 #' @param ps_equal_sign equal sign to be used when a variable name is given
 #' @param ps_env math environment in which matrix should be presented
+#' @param ps_out_format output format produced
 #' @return String containing bmatrix representation of given vector
 #'
 #' @examples
@@ -163,15 +184,20 @@ bcolumn_vector <- function(pvec, ps_name = NULL, ps_equal_sign = '=', ps_env = N
 #' }
 #'
 #' @export brow_vector
-brow_vector <- function(pvec, ps_name = NULL, ps_equal_sign = '=', ps_env = NULL){
+brow_vector <- function(pvec,
+                        ps_name = NULL,
+                        ps_equal_sign = '=',
+                        ps_env = NULL,
+                        ps_out_format = 'html'){
   # check type of pmat
   if (! is.vector(pvec))
     stop(" *** ERROR: specified pmat is not a vector")
   # Use bmatrix as output type
-  return(rmd_vec_mat(pmat          = as.matrix(pvec, nrow = 1),
+  return(rmd_vec_mat(pmat          = matrix(pvec, nrow = 1),
                      ps_name       = ps_name,
                      ps_equal_sign = ps_equal_sign,
                      ps_env        = ps_env,
-                     ps_out_type   = 'bmatrix'))
+                     ps_out_type   = 'bmatrix',
+                     ps_out_format = ps_out_format))
 }